+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 103:
 
 * Boost test matrix fixes
 * Tidy up allocator usage
+* Example HTTP server fixes
 
 --------------------------------------------------------------------------------